requirements.txt
# Google Gemini API (LLM için)
google-generativeai>=0.8.0
langchain>=0.1.0
langchain-google-genai>=0.0.6

# LangChain Core ve Community Modülleri
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-chroma>=0.1.0
langchain-text-splitters>=0.0.1

# Vektör Veritabanı
chromadb>=0.4.0

# Yerel Embedding Modeli
sentence-transformers>=2.2.0
torch>=2.0.0

# Veri İşleme
datasets>=2.14.0
pandas>=2.0.0
numpy>=1.24.0

# Ortam Değişkenleri
python-dotenv>=1.0.0

# Web Arayüzü
streamlit>=1.28.0
